knowledge-graph/sync/requirements.txt
aiohttp>=3.9.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
